Refs #124. Moved property hierarchy testing to its own module
